requirements.txt
asgiref==3.6.0
cx-Oracle==8.3.0
Django==3.2.18
django-environ==0.9.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.2
typing-extensions==4.5.0
